apply patch from Christophe Gisquet <christophe dot gisquet at free.fr> with some modifications -don't use access with X_OK, it won't work on Vista http://lists.samba.org/archive/ccache/2006q3/000242.html
git-svn-id: file:///srv/svn/reactos/trunk@32587 97493ccd-5924-5043-b1f5-66cb403b36ce

+++ tools/RosBE/RosBE-Windows/Tools/ccache-2.4/hash.c
@@ -35,7 +35,7 @@ void hash_start(void)
 
 void hash_string(const char *s)
 {
-	hash_buffer(s, strlen(s));
+	hash_buffer(s, (int)strlen(s));
 }
 
 void hash_int(int x)
